STOCKS-37: Handle versions correctly
Check version transactionally
Remove business manager
Move common classes

+++ server/src/main/java/de/njsm/stocks/server/v2/db/DatabaseHandler.java
@@ -1,11 +1,9 @@
 package de.njsm.stocks.server.v2.db;
 
-import de.njsm.stocks.server.v2.business.BusinessManager;
 import de.njsm.stocks.server.v2.business.StatusCode;
 import de.njsm.stocks.server.v2.business.data.Food;
 import fj.data.Validation;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
+import org.jooq.DSLContext;
 import org.jooq.Record;
 import org.jooq.Result;
 import org.jooq.types.UInteger;
@@ -17,8 +15,6 @@ import static de.njsm.stocks.server.v2.db.jooq.Tables.FOOD;
 
 public class DatabaseHandler extends FailSafeDatabaseHandler {
 
-    private static final Logger LOG = LogManager.getLogger(BusinessManager.class);
-
     public DatabaseHandler(String url,
                            String username,
                            String password,
@@ -27,15 +23,16 @@ public class DatabaseHandler extends FailSafeDatabaseHandler {
     }
 
     public StatusCode addFood(String name) {
-        return runOperation(context -> {
+        return runCommand(context -> {
            context.insertInto(FOOD, FOOD.NAME, FOOD.VERSION)
                    .values(name, UInteger.valueOf(1))
                    .execute();
+           return StatusCode.SUCCESS;
         });
     }
 
     public Validation<StatusCode, List<Food>> getFood() {
-        return runOperation(context -> {
+        return runQuery(context -> {
             Result<Record> cursor = context
                     .select()
                     .from(FOOD)
@@ -52,41 +49,50 @@ public class DatabaseHandler extends FailSafeDatabaseHandler {
         });
     }
 
-    public Validation<StatusCode, Food> getFood(int id) {
-        return runOperation(context -> {
-            Result<Record> cursor = context
-                    .select()
-                    .from(FOOD)
-                    .where(FOOD.ID.eq(UInteger.valueOf(id)))
-                    .fetch();
+    public StatusCode renameFood(int id, int version, String newName) {
+        return runCommand(context -> {
+            if (isMissing(id, context))
+                return StatusCode.NOT_FOUND;
 
-            if (cursor.isEmpty()) {
-                LOG.info("Unknown food id " + id + "demanded");
-                return Validation.fail(StatusCode.NOT_FOUND);
+            int changedItems = context.update(FOOD)
+                    .set(FOOD.NAME, newName)
+                    .where(FOOD.ID.eq(UInteger.valueOf(id))
+                            .and(FOOD.VERSION.eq(UInteger.valueOf(version))))
+                    .execute();
+
+            if (changedItems == 1) {
+                return StatusCode.SUCCESS;
             } else {
-                Food item = new Food();
-                Record r = cursor.get(0);
-                item.id = r.getValue(FOOD.ID).intValue();
-                item.name = r.getValue(FOOD.NAME);
-                item.version = r.getValue(FOOD.VERSION).intValue();
-                return Validation.success(item);
+                return StatusCode.INVALID_DATA_VERSION;
             }
         });
     }
 
-    public StatusCode renameFood(int id, String newName) {
-        return runOperation(context -> {
-            context.update(FOOD)
-                    .set(FOOD.NAME, newName)
-                    .where(FOOD.ID.eq(UInteger.valueOf(id)));
+    public StatusCode deleteFood(int id, int version) {
+        return runCommand(context -> {
+            if (isMissing(id, context))
+                return StatusCode.NOT_FOUND;
+
+            int changedItems = context.deleteFrom(FOOD)
+                    .where(FOOD.ID.eq(UInteger.valueOf(id))
+                            .and(FOOD.VERSION.eq(UInteger.valueOf(version))))
+                    .execute();
+
+            if (changedItems == 1) {
+                return StatusCode.SUCCESS;
+            } else {
+                return StatusCode.INVALID_DATA_VERSION;
+            }
         });
     }
 
-    public StatusCode deleteFood(int id) {
-        return runOperation(context -> {
-            context.deleteFrom(FOOD)
-                    .where(FOOD.ID.eq(UInteger.valueOf(id)))
-                    .execute();
-        });
+    private boolean isMissing(int id, DSLContext context) {
+        int count = context.select()
+                .from(FOOD)
+                .where(FOOD.ID.eq(UInteger.valueOf(id)))
+                .fetch()
+                .size();
+
+        return count == 0;
     }
 }
